+++ src/APIM_ARMTemplate/apimtemplate/Common/Templates/Abstractions/TemplateObjectParameterProperties.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace apimtemplate.Common.Templates.Abstractions
+{
+    public class TemplateObjectParameterProperties : TemplateParameterProperties
+    {
+        public new object value { get; set; }
+    }
+}
