+++ src/APIM_ARMTemplate/apimtemplate/Common/Templates/Abstractions/TemplateParameterMetadata.cs
@@ -1,0 +1,7 @@
+﻿namespace apimtemplate.Common.Templates.Abstractions
+{
+    public class TemplateParameterMetadata
+    {
+        public string description { get; set; }
+    }
+}
