+++ src/APIM_ARMTemplate/apimtemplate/Common/Templates/Abstractions/TemplateParameterProperties.cs
@@ -1,0 +1,11 @@
+﻿namespace apimtemplate.Common.Templates.Abstractions
+{
+    public class TemplateParameterProperties
+    {
+        public string type { get; set; }
+        public TemplateParameterMetadata metadata { get; set; }
+        public string[] allowedValues { get; set; }
+        public string defaultValue { get; set; }
+        public string value { get; set; }
+    }
+}
